update deploy scripts

+++ verify_api/deploy/before_install.sh
@@ -1,2 +1,8 @@
-rm -rf /home/ubuntu/app/
+if [ -d /home/ubuntu/app/ ]; then
+    cd /home/ubuntu/app/
+    # stop server
+    npm run server:stop
+    # remove directory
+    rm -rf /home/ubuntu/app/
+fi
 mkdir /home/ubuntu/app/
